pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.4.0",
    "flask>=3.1.0",
    "google-auth>=2.37.0",
    "gspread>=6.1.4",
    "python-dotenv>=1.0.1",
]
